Handle actions part of welcome event
