+++ app.json
@@ -19,5 +19,5 @@
         "IPostLivechatRoomClosed",
         "IUIKitLivechatInteractionHandler"
     ],
-    "commitHash": "f6d978e0fa23574ebf646ac0b1c371b15494127f"
+    "commitHash": "47a27ed35ee6fe974a5043fe82d49b6866a63827"
 }
